+++ agent/src/main/java/com/microsoft/hydralab/agent/scheduled/ScheduledDeviceControlTasks.java
@@ -49,6 +49,8 @@ public class ScheduledDeviceControlTasks {
         clearFile(appOptions.getTestPackageLocation());
         logger.info("schedule clean test result");
         clearFile(appOptions.getTestCaseResultLocation());
+        logger.info("schedule clean error output");
+        clearFile(appOptions.getErrorStorageLocation());
     }
 
     public void clearFile(String folderPath) {
